Fixed the rotation by making it a quaternion instead
