+++ Assets/Scenes/CannonScene.unity
@@ -135,6 +135,7 @@ GameObject:
   - component: {fileID: 1352707903}
   - component: {fileID: 1352707902}
   - component: {fileID: 1352707901}
+  - component: {fileID: 1352707904}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -216,6 +217,50 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1352707904
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1352707900}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a79441f348de89743a2939f4d699eac1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_RenderShadows: 1
+  m_RequiresDepthTextureOption: 2
+  m_RequiresOpaqueTextureOption: 2
+  m_CameraType: 0
+  m_Cameras: []
+  m_RendererIndex: -1
+  m_VolumeLayerMask:
+    serializedVersion: 2
+    m_Bits: 1
+  m_VolumeTrigger: {fileID: 0}
+  m_VolumeFrameworkUpdateModeOption: 2
+  m_RenderPostProcessing: 0
+  m_Antialiasing: 0
+  m_AntialiasingQuality: 2
+  m_StopNaN: 0
+  m_Dithering: 0
+  m_ClearDepth: 1
+  m_AllowXRRendering: 1
+  m_AllowHDROutput: 1
+  m_UseScreenCoordOverride: 0
+  m_ScreenSizeOverride: {x: 0, y: 0, z: 0, w: 0}
+  m_ScreenCoordScaleBias: {x: 0, y: 0, z: 0, w: 0}
+  m_RequiresDepthTexture: 0
+  m_RequiresColorTexture: 0
+  m_Version: 2
+  m_TaaSettings:
+    m_Quality: 3
+    m_FrameInfluence: 0.1
+    m_JitterScale: 1
+    m_MipBias: 0
+    m_VarianceClampScale: 0.9
+    m_ContrastAdaptiveSharpening: 0
 --- !u!1 &1961333474
 GameObject:
   m_ObjectHideFlags: 0
@@ -343,8 +388,8 @@ Transform:
   m_GameObject: {fileID: 1961333474}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.06, y: -0.91, z: 0}
-  m_LocalScale: {x: 17.5977, y: 1, z: 1}
+  m_LocalPosition: {x: 0.06, y: -0.87, z: 0}
+  m_LocalScale: {x: 100, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -359,8 +404,16 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 1183189936547476379, guid: 96852da69c05945418271ec246e03d10, type: 3}
       propertyPath: a
-      value: 1
+      value: 0.66
       objectReference: {fileID: 0}
+    - target: {fileID: 1183189936547476379, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: _a
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 1183189936547476379, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: barrelRotationPivot
+      value: 
+      objectReference: {fileID: 545134947}
     - target: {fileID: 1183189936547476379, guid: 96852da69c05945418271ec246e03d10, type: 3}
       propertyPath: _barrelRotationPivot
       value: 
@@ -368,10 +421,30 @@ PrefabInstance:
     - target: {fileID: 1968205763515253657, guid: 96852da69c05945418271ec246e03d10, type: 3}
       propertyPath: prefabToPool
       value: 
-      objectReference: {fileID: 970915876847491512, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
+      objectReference: {fileID: 8489824262079630986, guid: d58c385af0c1e264ea9dbd2cc8b91cb9, type: 3}
+    - target: {fileID: 1968205763515253657, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: shootPosition
+      value: 
+      objectReference: {fileID: 545134947}
     - target: {fileID: 3145879125921422353, guid: 96852da69c05945418271ec246e03d10, type: 3}
       propertyPath: m_Name
       value: Cannon
+      objectReference: {fileID: 0}
+    - target: {fileID: 5340278274252750456, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.9577603
+      objectReference: {fileID: 0}
+    - target: {fileID: 5340278274252750456, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5340278274252750456, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5340278274252750456, guid: 96852da69c05945418271ec246e03d10, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0.2875679
       objectReference: {fileID: 0}
     - target: {fileID: 7293545863962168998, guid: 96852da69c05945418271ec246e03d10, type: 3}
       propertyPath: m_Name
@@ -422,68 +495,10 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 96852da69c05945418271ec246e03d10, type: 3}
---- !u!1001 &2252181368041125483
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 970915876847491512, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_Name
-      value: Projectile
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 1.82
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 3.01
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2253174784473293930, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: cde023628dcb81b4b863ed3f98471a93, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
   m_Roots:
   - {fileID: 1352707903}
   - {fileID: 1808862894260923872}
-  - {fileID: 2252181368041125483}
   - {fileID: 1961333477}
