+++ Assets/Scripts/Cannon/CannonAngleController.cs
@@ -1,31 +1,50 @@
-using System.Collections;
-using System.Collections.Generic;
+using TypedUnityEvent;
 using UnityEngine;
-using UnityEngine.Events;
 
-public class CannonAngleController : MonoBehaviour
+namespace Cannon
 {
-    [SerializeField]
-    private GameObject _barrelRotationPivot;
-
-    [SerializeField]
-    private float a; 
-
-    private void FixedUpdate()
+    public class CannonAngleController : MonoBehaviour
     {
-        Rotate();
-    }
+        [SerializeField] private GameObject barrelRotationPivot;
+        [SerializeField] private float a;
+        [SerializeField] private FloatEvent OnAngleChange = new();
 
-    // Rotate the barrel to the specified angle
-    public void Rotate()
-    {
-        _barrelRotationPivot.transform.rotation = Quaternion.Euler(0f, 0f, GetAngle(a));
-    }
+        public float A
+        {
+            get { return a; }
+            set
+            {
+                a = value;
+                Rotate();
+            }
+        }
 
-    // Calculate the angle in degrees
-    private float GetAngle(float a)
-    {
-        float x = 1;
-        return Mathf.Atan2(a * x, x) * Mathf.Rad2Deg;
+
+        /// <summary>
+        /// Rotate the barrel based on the value of _a
+        /// </summary>
+        private void Rotate()
+        {
+            var newAngle = GetAngle(a);
+            barrelRotationPivot.transform.rotation = Quaternion.Euler(0f, 0f, newAngle);
+            OnAngleChange.Invoke(newAngle);
+        }
+
+        /// <summary>
+        /// Calculate the angle in degrees
+        /// </summary>
+        /// <param name="a">The Y coordinate of linear function</param>
+        /// <returns>The calculated angle in degrees</returns>
+        private float GetAngle(float y)
+        {
+            float x = 1;
+            return Mathf.Atan2(y * x, x) * Mathf.Rad2Deg;
+        }
+
+        private void OnValidate()
+        {
+            if (barrelRotationPivot != null)
+                Rotate();
+        }
     }
 }
