+++ Assets/Scripts/Cannon/CannonProjectileController.cs
@@ -1,95 +1,100 @@
 using UnityEngine;
 using UnityEngine.Events;
 using System.Collections.Generic;
-using Unity.Mathematics;
+using Projectile;
 
-public class CannonProjectileController : MonoBehaviour
+namespace Cannon
 {
-
-    [SerializeField] private GameObject prefabToPool;
-
-    [SerializeField] private int amountToPool = 20;
-
-    [SerializeField] private UnityEvent OnCannonFire = new();
-
-    private List<ProjectileScript> _pooledProjectiles = new();
-
-    // Start is called before the first frame update
-    private void Start()
+    public class CannonProjectileController : MonoBehaviour
     {
-        CreatePooledProjectiles();
-    }
+        [SerializeField] private GameObject shootPosition;
 
-    private void OnValidate()
-    {
-        if (Application.isPlaying)
+        [SerializeField] private GameObject prefabToPool;
+
+        [SerializeField] private int amountToPool = 20;
+
+        [SerializeField] private UnityEvent OnCannonFire = new();
+
+        private List<ProjectileScript> _pooledProjectiles = new();
+
+        // Start is called before the first frame update
+        private void Start()
+        {
             CreatePooledProjectiles();
-    }
-
-
-    /// <summary>
-    /// Empties the current pooled object list and creates a new pool
-    /// </summary>
-    private void CreatePooledProjectiles()
-    {
-        if (prefabToPool.GetComponent<ProjectileScript>() == null)
-        {
-            Debug.LogWarning($"Prefab {prefabToPool.name} doesn't have a ProjectileScript!");
-            return;
         }
 
-        // Checks if there are already pooled objects and destroys exisitng pooled gameObjects
-        if (_pooledProjectiles != null && _pooledProjectiles.Count > 0)
+        private void OnValidate()
         {
-            for (var i = 0; i < _pooledProjectiles.Count; i++)
+            if (Application.isPlaying)
+                CreatePooledProjectiles();
+        }
+
+
+        /// <summary>
+        /// Empties the current pooled object list and creates a new pool
+        /// </summary>
+        private void CreatePooledProjectiles()
+        {
+            if (prefabToPool.GetComponent<ProjectileScript>() == null)
             {
-                Destroy(_pooledProjectiles[i].gameObject);
+                Debug.LogWarning($"Prefab {prefabToPool.name} doesn't have a ProjectileScript!");
+                return;
+            }
+
+            // Checks if there are already pooled objects and destroys exisitng pooled gameObjects
+            if (_pooledProjectiles != null && _pooledProjectiles.Count > 0)
+            {
+                for (var i = 0; i < _pooledProjectiles.Count; i++)
+                {
+                    Destroy(_pooledProjectiles[i].gameObject);
+                }
+            }
+
+            // Creates a new list of pooled objects
+            _pooledProjectiles = new();
+            GameObject tmp;
+            for (var i = 0; i < amountToPool; i++)
+            {
+                tmp = Instantiate(prefabToPool);
+                tmp.SetActive(false);
+                _pooledProjectiles.Add(tmp.GetComponent<ProjectileScript>());
             }
         }
 
-        // Creates a new list of pooled objects
-        _pooledProjectiles = new();
-        GameObject tmp;
-        for (var i = 0; i < amountToPool; i++)
+        /// <summary>
+        /// Returns a pooled projectile if available, otherwise returns null
+        /// </summary>
+        private ProjectileScript GetPooledProjectile()
         {
-            tmp = Instantiate(prefabToPool);
-            tmp.SetActive(false);
-            _pooledProjectiles.Add(tmp.GetComponent<ProjectileScript>());
-        }
-    }
-
-    /// <summary>
-    /// Returns a pooled projectile if available, otherwise returns null
-    /// </summary>
-    private ProjectileScript GetPooledProjectile()
-    {
-        for (var i = 0; i < amountToPool; i++)
-        {
-            if (!_pooledProjectiles[i].gameObject.activeInHierarchy)
+            for (var i = 0; i < amountToPool; i++)
             {
-                return _pooledProjectiles[i];
+                if (!_pooledProjectiles[i].gameObject.activeInHierarchy)
+                {
+                    return _pooledProjectiles[i];
+                }
             }
+            return null;
         }
-        return null;
-    }
 
-    /// <summary>
-    /// Activates a pooled projectile and invokes the cannon fired event 
-    /// </summary>
-    public void ShootProjectile()
-    {
-        ProjectileScript projectile = GetPooledProjectile();
-        if (projectile)
+        /// <summary>
+        /// Activates a pooled projectile and invokes the cannon fired event 
+        /// </summary>
+        public void ShootProjectile()
         {
-            projectile.transform.position = transform.position;
-            projectile.gameObject.SetActive(true);
-            projectile.Shoot(Vector2.right);
+            ProjectileScript projectile = GetPooledProjectile();
+            if (projectile)
+            {
+                projectile.transform.position = transform.position;
+                projectile.gameObject.SetActive(true);
 
-            OnCannonFire.Invoke();
-        }
-        else
-        {
-            Debug.LogWarning($"Not enough pooled projectiles! Consider increasing the current amount: {amountToPool}");
+                projectile.Shoot(shootPosition.transform.rotation);
+
+                OnCannonFire.Invoke();
+            }
+            else
+            {
+                Debug.LogWarning($"Not enough pooled projectiles! Consider increasing the current amount: {amountToPool}");
+            }
         }
     }
 }
