+++ Assets/Scripts/Cannon/Editor/CannonProjectileControllerEditor.cs
@@ -1,3 +1,4 @@
+using Cannon;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEditor;
